feat: create service for interaction with user

+++ src/Lab5/Lab5Project.Application.Models/Users/User.cs
@@ -1,3 +1,3 @@
 namespace Lab5Project.Application.Models.Users;
 
-public record User(long AccountNumber, string Username, long Pin, decimal Amount);
+public record User(long AccountNumber, string Username, int Pin, decimal Amount);
